// src/client/types.gen.ts
export type ProjectResource = {
    id: number;
    title: string;
    description: string | null;
    is_completed: boolean;
    completed_at: number | null;
    completion_percentage: number;
    parent_id: number | null;
    created_at: string;
    updated_at: string;
    /**
     * 'user_id' => $this->user_id,
     * 'tag_id' => $this->tag_id,
     */
    tasks?: Array<ProjectTaskResource>;
    tag?: TagResource;
    subprojects?: Array<ProjectResource>;
};

/**
 * ProjectTaskRequest
 */
export type ProjectTaskRequest = {
    title: string;
    description?: string | null;
    is_completed?: boolean;
    completed_at?: string | null;
    project_id: number;
};

/**
 * ProjectTaskResource
 */
export type ProjectTaskResource = {
    id: number;
    title: string;
    description: string | null;
    is_completed: boolean;
    completed_at: number | null;
    created_at: string;
    updated_at: string;
    project_id: number;
    project?: ProjectResource;
};

/**
 * TagRequest
 */
export type TagRequest = {
    name: string;
    color: string;
    tags_type: 'school' | 'reminder' | 'project' | 'task';
};

/**
 * TagResource
 */
export type TagResource = {
    id: number;
    name: string;
    color: string;
    tags_type: string;
    created_at: string;
    updated_at: string;
};
// ...
export type ProjectsIndexResponses = {
    200: {
        completed: Array<ProjectResource>;
        uncompleted: Array<ProjectResource>;
    };
};

export type ProjectsIndexResponse = ProjectsIndexResponses[keyof ProjectsIndexResponses];
// ...
export type ProjectsShowResponses = {
    /**
     * `ProjectResource`
     */
    200: {
        data: ProjectResource;
    };
};

export type ProjectsShowResponse = ProjectsShowResponses[keyof ProjectsShowResponses];